# Conflicts:
#	java/filament/src/main/cpp/RenderableManager.cpp
#	java/filament/src/main/cpp/Renderer.cpp
#	java/filament/src/main/java/io/github/erkko68/filament/jni/RenderableManager.java
#	java/filament/src/main/java/io/github/erkko68/filament/jni/Renderer.java
#	kotlin/filament/src/androidMain/kotlin/io/github/erkko68/filament/Renderer.android.kt
#	kotlin/filament/src/jvmMain/kotlin/io/github/erkko68/filament/RenderableManager.jvm.kt
#	kotlin/filament/src/jvmMain/kotlin/io/github/erkko68/filament/Renderer.jvm.kt
#	scripts/README.md

+++ c/filament/c/Renderer.h
@@ -19,7 +19,7 @@ typedef struct FilaRendererFrameRateOptions {
 } FilaRendererFrameRateOptions;
 
 typedef struct FilaRendererClearOptions {
-    float clearColor[4];
+    double clearColor[4];
     bool clear;
     bool discard;
 } FilaRendererClearOptions;
